feat: Extract MCP tool handlers to separate modules (Phase 3)
- Created handlers directory with domain-specific handler modules:
  - projects.ts: Project listing handler
  - issues.ts: Issue retrieval handler with detailed mapping
  - metrics.ts: Metrics handler
  - system.ts: Health, status, and ping handlers
  - measures.ts: Component measures and history handlers
  - quality-gates.ts: Quality gate handlers
  - source-code.ts: Source code and SCM blame handlers
  - hotspots.ts: Security hotspot handlers

- Moved client factory logic to utils/client-factory.ts:
  - Moved getDefaultClient() from index.ts
  - Moved resetDefaultClient() from index.ts
  - Maintains lazy initialization for environment validation

- Updated index.ts to import handlers instead of defining them:
  - Removed handler function implementations
  - Added imports from handlers/index.js
  - Re-exports handlers for backward compatibility
  - Cleaned up unused type imports

All tests passing, TypeScript checks pass, and code is properly formatted.

Part of #104: Refactor codebase into smaller, focused modules
