+++ src/index.ts
@@ -3,26 +3,35 @@
 import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
 import { StdioServerTransport } from '@modelcontextprotocol/sdk/server/stdio.js';
 import {
-  PaginationParams,
   ISonarQubeClient,
-  IssuesParams,
-  SonarQubeProject,
-  SonarQubeIssue,
-  ComponentMeasuresParams,
-  ComponentsMeasuresParams,
-  MeasuresHistoryParams,
-  ProjectQualityGateParams,
-  SourceCodeParams,
-  ScmBlameParams,
+  createSonarQubeClientFromEnv,
   HotspotSearchParams,
   HotspotStatusUpdateParams,
-  createSonarQubeClientFromEnv,
 } from './sonarqube.js';
 import { z } from 'zod';
 import { createLogger } from './utils/logger.js';
 import { nullToUndefined, stringToNumberTransform } from './utils/transforms.js';
 import { mapToSonarQubeParams } from './utils/parameter-mappers.js';
-import { validateEnvironmentVariables } from './utils/client-factory.js';
+import { validateEnvironmentVariables, resetDefaultClient } from './utils/client-factory.js';
+import {
+  handleSonarQubeProjects,
+  handleSonarQubeGetIssues,
+  handleSonarQubeGetMetrics,
+  handleSonarQubeGetHealth,
+  handleSonarQubeGetStatus,
+  handleSonarQubePing,
+  handleSonarQubeComponentMeasures,
+  handleSonarQubeComponentsMeasures,
+  handleSonarQubeMeasuresHistory,
+  handleSonarQubeListQualityGates,
+  handleSonarQubeGetQualityGate,
+  handleSonarQubeQualityGateStatus,
+  handleSonarQubeGetSourceCode,
+  handleSonarQubeGetScmBlame,
+  handleSonarQubeHotspots,
+  handleSonarQubeHotspot,
+  handleSonarQubeUpdateHotspotStatus,
+} from './handlers/index.js';
 
 const logger = createLogger('index');
 
@@ -62,466 +71,29 @@ export const createDefaultClient = (): ISonarQubeClient => {
   return client;
 };
 
-// Default client instance for backward compatibility
-// Created lazily to allow environment variable validation at runtime
-let defaultClient: ISonarQubeClient | null = null;
+// Re-export resetDefaultClient for backward compatibility
+export { resetDefaultClient };
 
-const getDefaultClient = (): ISonarQubeClient => {
-  defaultClient ??= createDefaultClient();
-  return defaultClient;
-};
-
-// Export for testing purposes
-export const resetDefaultClient = () => {
-  defaultClient = null;
-};
-
-/**
- * Fetches and returns a list of all SonarQube projects
- * @param params Parameters for listing projects, including pagination and organization
- * @param client Optional SonarQube client instance
- * @returns A response containing the list of projects with their details
- * @throws Error if no authentication environment variables are set (SONARQUBE_TOKEN, SONARQUBE_USERNAME/PASSWORD, or SONARQUBE_PASSCODE)
- */
-export async function handleSonarQubeProjects(
-  params: {
-    page?: number | null;
-    page_size?: number | null;
-  },
-  client: ISonarQubeClient = getDefaultClient()
-) {
-  logger.debug('Handling SonarQube projects request', params);
-
-  const projectsParams: PaginationParams = {
-    page: nullToUndefined(params.page),
-    pageSize: nullToUndefined(params.page_size),
-  };
-
-  try {
-    const result = await client.listProjects(projectsParams);
-    logger.info('Successfully retrieved projects', { count: result.projects.length });
-    return {
-      content: [
-        {
-          type: 'text' as const,
-          text: JSON.stringify({
-            projects: result.projects.map((project: SonarQubeProject) => ({
-              key: project.key,
-              name: project.name,
-              qualifier: project.qualifier,
-              visibility: project.visibility,
-              lastAnalysisDate: project.lastAnalysisDate,
-              revision: project.revision,
-              managed: project.managed,
-            })),
-            paging: result.paging,
-          }),
-        },
-      ],
-    };
-  } catch (error) {
-    logger.error('Failed to retrieve SonarQube projects', error);
-    throw error;
-  }
-}
-
-/**
- * Fetches and returns issues from a specified SonarQube project
- * @param params Parameters for fetching issues, including project key, severity, and pagination
- * @param client Optional SonarQube client instance
- * @returns A response containing the list of issues with their details
- * @throws Error if no authentication environment variables are set (SONARQUBE_TOKEN, SONARQUBE_USERNAME/PASSWORD, or SONARQUBE_PASSCODE)
- */
-export async function handleSonarQubeGetIssues(
-  params: IssuesParams,
-  client: ISonarQubeClient = getDefaultClient()
-) {
-  logger.debug('Handling SonarQube issues request', { projectKey: params.projectKey });
-
-  try {
-    const result = await client.getIssues(params);
-    logger.info('Successfully retrieved issues', {
-      projectKey: params.projectKey,
-      count: result.issues.length,
-    });
-
-    return {
-      content: [
-        {
-          type: 'text' as const,
-          text: JSON.stringify({
-            issues: result.issues.map((issue: SonarQubeIssue) => ({
-              key: issue.key,
-              rule: issue.rule,
-              severity: issue.severity,
-              component: issue.component,
-              project: issue.project,
-              line: issue.line,
-              status: issue.status,
-              issueStatus: issue.issueStatus,
-              message: issue.message,
-              messageFormattings: issue.messageFormattings,
-              effort: issue.effort,
-              debt: issue.debt,
-              author: issue.author,
-              tags: issue.tags,
-              creationDate: issue.creationDate,
-              updateDate: issue.updateDate,
-              type: issue.type,
-              cleanCodeAttribute: issue.cleanCodeAttribute,
-              cleanCodeAttributeCategory: issue.cleanCodeAttributeCategory,
-              prioritizedRule: issue.prioritizedRule,
-              impacts: issue.impacts,
-              textRange: issue.textRange,
-              comments: issue.comments,
-              transitions: issue.transitions,
-              actions: issue.actions,
-              flows: issue.flows,
-              quickFixAvailable: issue.quickFixAvailable,
-              ruleDescriptionContextKey: issue.ruleDescriptionContextKey,
-              codeVariants: issue.codeVariants,
-              hash: issue.hash,
-            })),
-            components: result.components,
-            rules: result.rules,
-            users: result.users,
-            facets: result.facets,
-            paging: result.paging,
-          }),
-        },
-      ],
-    };
-  } catch (error) {
-    logger.error('Failed to retrieve SonarQube issues', error);
-    throw error;
-  }
-}
-
-/**
- * Handler for getting SonarQube metrics
- * @param params Parameters for the metrics request
- * @param client Optional SonarQube client instance
- * @returns Promise with the metrics result
- */
-export async function handleSonarQubeGetMetrics(
-  params: PaginationParams,
-  client: ISonarQubeClient = getDefaultClient()
-) {
-  const result = await client.getMetrics(params);
-
-  // Create a properly structured response matching the expected format
-  const response = {
-    metrics: result.metrics ?? [],
-    paging: result.paging ?? {
-      pageIndex: params.page ?? 1,
-      pageSize: params.pageSize ?? 100,
-      total: (result.metrics ?? []).length,
-    },
-  };
-
-  return {
-    content: [
-      {
-        type: 'text' as const,
-        text: JSON.stringify(response),
-      },
-    ],
-  };
-}
-
-/**
- * Handler for getting SonarQube system health status
- * @param client Optional SonarQube client instance
- * @returns Promise with the health status result
- */
-export async function handleSonarQubeGetHealth(client: ISonarQubeClient = getDefaultClient()) {
-  const result = await client.getHealth();
-
-  return {
-    content: [
-      {
-        type: 'text' as const,
-        text: JSON.stringify(result),
-      },
-    ],
-  };
-}
-
-/**
- * Handler for getting SonarQube system status
- * @param client Optional SonarQube client instance
- * @returns Promise with the system status result
- */
-export async function handleSonarQubeGetStatus(client: ISonarQubeClient = getDefaultClient()) {
-  const result = await client.getStatus();
-
-  return {
-    content: [
-      {
-        type: 'text' as const,
-        text: JSON.stringify(result),
-      },
-    ],
-  };
-}
-
-/**
- * Handler for pinging SonarQube system
- * @param client Optional SonarQube client instance
- * @returns Promise with the ping result
- */
-export async function handleSonarQubePing(client: ISonarQubeClient = getDefaultClient()) {
-  const result = await client.ping();
-
-  return {
-    content: [
-      {
-        type: 'text' as const,
-        text: result,
-      },
-    ],
-  };
-}
-
-/**
- * Handler for getting measures for a specific component
- * @param params Parameters for the component measures request
- * @param client Optional SonarQube client instance
- * @returns Promise with the component measures result
- */
-export async function handleSonarQubeComponentMeasures(
-  params: ComponentMeasuresParams,
-  client: ISonarQubeClient = getDefaultClient()
-) {
-  const result = await client.getComponentMeasures(params);
-
-  return {
-    content: [
-      {
-        type: 'text' as const,
-        text: JSON.stringify(result),
-      },
-    ],
-  };
-}
-
-/**
- * Handler for getting measures for multiple components
- * @param params Parameters for the components measures request
- * @param client Optional SonarQube client instance
- * @returns Promise with the components measures result
- */
-export async function handleSonarQubeComponentsMeasures(
-  params: ComponentsMeasuresParams,
-  client: ISonarQubeClient = getDefaultClient()
-) {
-  const result = await client.getComponentsMeasures(params);
-
-  return {
-    content: [
-      {
-        type: 'text' as const,
-        text: JSON.stringify(result),
-      },
-    ],
-  };
-}
-
-/**
- * Handler for getting measures history for a component
- * @param params Parameters for the measures history request
- * @param client Optional SonarQube client instance
- * @returns Promise with the measures history result
- */
-export async function handleSonarQubeMeasuresHistory(
-  params: MeasuresHistoryParams,
-  client: ISonarQubeClient = getDefaultClient()
-) {
-  const result = await client.getMeasuresHistory(params);
-
-  return {
-    content: [
-      {
-        type: 'text' as const,
-        text: JSON.stringify(result),
-      },
-    ],
-  };
-}
-
-/**
- * Handler for listing quality gates
- * @param client Optional SonarQube client instance
- * @returns Promise with the list of quality gates
- */
-export async function handleSonarQubeListQualityGates(
-  client: ISonarQubeClient = getDefaultClient()
-) {
-  const result = await client.listQualityGates();
-
-  return {
-    content: [
-      {
-        type: 'text' as const,
-        text: JSON.stringify(result),
-      },
-    ],
-  };
-}
-
-/**
- * Handler for getting quality gate details
- * @param params Parameters for the quality gate request
- * @param client Optional SonarQube client instance
- * @returns Promise with the quality gate details
- */
-export async function handleSonarQubeGetQualityGate(
-  params: { id: string },
-  client: ISonarQubeClient = getDefaultClient()
-) {
-  const result = await client.getQualityGate(params.id);
-
-  return {
-    content: [
-      {
-        type: 'text' as const,
-        text: JSON.stringify(result),
-      },
-    ],
-  };
-}
-
-/**
- * Handler for getting project quality gate status
- * @param params Parameters for the project quality gate status request
- * @param client Optional SonarQube client instance
- * @returns Promise with the project quality gate status
- */
-export async function handleSonarQubeQualityGateStatus(
-  params: ProjectQualityGateParams,
-  client: ISonarQubeClient = getDefaultClient()
-) {
-  const result = await client.getProjectQualityGateStatus(params);
-
-  return {
-    content: [
-      {
-        type: 'text' as const,
-        text: JSON.stringify(result),
-      },
-    ],
-  };
-}
-
-/**
- * Handler for getting source code with issues
- * @param params Parameters for the source code request
- * @param client Optional SonarQube client instance
- * @returns Promise with the source code and annotations
- */
-export async function handleSonarQubeGetSourceCode(
-  params: SourceCodeParams,
-  client: ISonarQubeClient = getDefaultClient()
-) {
-  const result = await client.getSourceCode(params);
-
-  return {
-    content: [
-      {
-        type: 'text' as const,
-        text: JSON.stringify(result),
-      },
-    ],
-  };
-}
-
-/**
- * Handler for getting SCM blame information
- * @param params Parameters for the SCM blame request
- * @param client Optional SonarQube client instance
- * @returns Promise with the blame information
- */
-export async function handleSonarQubeGetScmBlame(
-  params: ScmBlameParams,
-  client: ISonarQubeClient = getDefaultClient()
-) {
-  const result = await client.getScmBlame(params);
-
-  return {
-    content: [
-      {
-        type: 'text' as const,
-        text: JSON.stringify(result),
-      },
-    ],
-  };
-}
-
-/**
- * Handler for searching security hotspots
- * @param params Parameters for the hotspot search
- * @param client Optional SonarQube client instance
- * @returns Promise with the hotspot search results
- */
-export async function handleSonarQubeHotspots(
-  params: HotspotSearchParams,
-  client: ISonarQubeClient = getDefaultClient()
-) {
-  const result = await client.hotspots(params);
-
-  return {
-    content: [
-      {
-        type: 'text' as const,
-        text: JSON.stringify(result),
-      },
-    ],
-  };
-}
-
-/**
- * Handler for getting hotspot details
- * @param hotspotKey The key of the hotspot
- * @param client Optional SonarQube client instance
- * @returns Promise with the hotspot details
- */
-export async function handleSonarQubeHotspot(
-  hotspotKey: string,
-  client: ISonarQubeClient = getDefaultClient()
-) {
-  const result = await client.hotspot(hotspotKey);
-
-  return {
-    content: [
-      {
-        type: 'text' as const,
-        text: JSON.stringify(result),
-      },
-    ],
-  };
-}
-
-/**
- * Handler for updating hotspot status
- * @param params Parameters for updating hotspot status
- * @param client Optional SonarQube client instance
- * @returns Promise that resolves when the update is complete
- */
-export async function handleSonarQubeUpdateHotspotStatus(
-  params: HotspotStatusUpdateParams,
-  client: ISonarQubeClient = getDefaultClient()
-) {
-  await client.updateHotspotStatus(params);
-
-  return {
-    content: [
-      {
-        type: 'text' as const,
-        text: 'Hotspot status updated successfully',
-      },
-    ],
-  };
-}
+// Re-export handlers for backward compatibility
+export {
+  handleSonarQubeProjects,
+  handleSonarQubeGetIssues,
+  handleSonarQubeGetMetrics,
+  handleSonarQubeGetHealth,
+  handleSonarQubeGetStatus,
+  handleSonarQubePing,
+  handleSonarQubeComponentMeasures,
+  handleSonarQubeComponentsMeasures,
+  handleSonarQubeMeasuresHistory,
+  handleSonarQubeListQualityGates,
+  handleSonarQubeGetQualityGate,
+  handleSonarQubeQualityGateStatus,
+  handleSonarQubeGetSourceCode,
+  handleSonarQubeGetScmBlame,
+  handleSonarQubeHotspots,
+  handleSonarQubeHotspot,
+  handleSonarQubeUpdateHotspotStatus,
+} from './handlers/index.js';
 
 // Define SonarQube severity schema for validation
 const severitySchema = z
